fix: fixed delay on delete column and private routing

+++ src/api/tasks/deleteTaskById.ts
@@ -3,13 +3,17 @@ import { createClient } from "../../utils/supabase/component"
 export default async function deleteTaskById(id: string) {
     try {
         const supabase = createClient();
+        const userId = (await supabase.auth.getSession()).data.session?.user.id
+
         const { error } = await supabase
             .from('task')
             .delete()
             .eq('id', id)
+            .eq("user_id", userId)
         
         if (error) {
             console.error(error)
+            return;
         }
     }
